content/html/content/test/test_bug386496.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=386496
-->
<head>
  <title>Test for Bug 386496</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=386496">Mozilla Bug 386496</a>
<p id="display"></p>
<div id="content">
  <div id="frameContent">
    <a id="a" href="http://a.invalid/">Link</a>
  </div>
  <iframe id="testIframe"></iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 386496 **/

var frame = document.getElementById("testIframe");

function finishTest() {
  is(frame.contentWindow.location, "about:blank");
  SimpleTest.finish();
}

function testDesignMode() {
  frame.style.display = 'block';
  var frameContent = frame.contentDocument.adoptNode(document.getElementById("frameContent"));
  frame.contentDocument.body.appendChild(frameContent);

  frame.contentDocument.designMode = "on";

  netscape.security.PrivilegeManager.enablePrivilege('UniversalBrowserWrite');
  var event = frame.contentDocument.createEvent('MouseEvent');
  event.initMouseEvent("click", true, true, window, 1, 0, 0, 0, 0, false, false,
                       false, false, 0, null);
  frame.contentDocument.getElementById("a").dispatchEvent(event);

  setTimeout(finishTest, 100);
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(testDesignMode);

</script>
</pre>
</body>
</html>
